+++ microservices/order-server/src/main/java/com/anup/ecommerce/entity/OrderLine.java
@@ -12,11 +12,11 @@ import lombok.*;
 public class OrderLine {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "order_id") // optional since JPA generates order + _ + id
     private Order order;
-    private Long productId;
+    private Integer productId;
     private double quantity;
 }
